generate-faiss-reference-index/requirements.txt
boto3==1.35.41
numpy==1.25.2
faiss-cpu==1.7.2